Profile graphql returns score (#375)
