+++ app/graphql/schema.graphql
@@ -211,6 +211,12 @@ type Profile implements Node & RulesPreload {
     """
     systemId: String
   ): Int!
+  score(
+    """
+    Latest TestResult score for this system and profile
+    """
+    systemId: String
+  ): Float!
   totalHostCount: Int!
 }
